backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
python-jose[cryptography]==3.3.0
pydantic==2.5.3
python-dotenv==1.0.0
PyJWT==2.8.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
